Create v1
The goal for this programming project is to create a simple 2D predator-prey simulation.  In this simulation the prey are ants"X" and the predators are doodlebugs"o". These critters live in a world composed of a 20x20 grid of cells. Only one critter may occupy a cell at a time. The grid is enclosed, so a critter is not allowed to move off the edges of the world. Time is simulated in time steps. Each critter performs some action every time step.

The ants behave according to the following rules:

- Move. Every time step, randomly try to move up, down, left or right. If the neighboring cell in the selected direction is occupied or would move the ant off the grid, then the ant stays in the current cell.

- Breed. If an ant survives for three time steps, then at the end of the third time step (i.e. after moving) the ant will breed. This is simulated by creating a new ant in an adjacent (up, down, left, or right) cell that is empty. This cell can be selected deterministically (e.g. always check left first, then up, etc.) or randomly. If there is no empty cell available then no breeding occurs. Once an offspring is produced an ant cannot produce an offspring until three more time steps have elapsed.

The doodlebugs behave according to the following rules:

- Move. Every time step, if there is an adjacent ant (up, down, left, or right) then the doodlebug will move to that cell and eat the ant. Otherwise the doodlebug moves according to the same rules as the ant. Note that a doodlebug cannot eat other doodlebugs.

- Breed. If a doodlebug survives for eight time steps, then at the end of the eighth time step it will spawn off a new doodlebug in the same manner as the ant.

- Starve. If a doodlebug has not eaten an ant within the last three time steps, then at the end of the third time step it will starve and die. The doodlebug should then be removed from the grid of cells. Note that ants don't starve, they only die by being eaten.

During one turn, all the doodlebugs should move/breed/starve before the ants move/breed.